+++ custom_typings/whatwg-url.d.ts
@@ -1,8 +1,22 @@
 declare module 'whatwg-url' {
+  export function parseURL(
+      input: string, options?: {baseURL?: URL, encodingOverride?: string}): URL|
+      null;
+
+  // https://url.spec.whatwg.org/#url-class
   export class URL {
+    constructor(url: string);
+
     protocol: string;
     pathname: string;
-
-    constructor(url: string);
+    scheme: string;
+    username: string;
+    password: string;
+    host: string;
+    port: number|null;
+    path: string[];
+    query: string|null;
+    fragment: string|null;
+    cannotBeABaseURL: boolean;
   }
 }