+++ src/test/static/javascript/module-with-named-import.js
@@ -1,1 +1,2 @@
 import * as test from 'test-package';
+import * as test2 from 'test-package/index';